+++ qutebrowser/browser/webkit/webkittab.py
@@ -646,8 +646,6 @@ class WebKitTab(browsertab.AbstractTab):
     @pyqtSlot(QUrl)
     def _on_url_changed(self, url):
         """Update title when URL has changed and no title is available."""
-        self._widget.apply_local_zoom_policy(url)
-        self._widget.apply_local_js_policy(url)
         super()._on_url_changed(url)
 
     def _install_event_filter(self):
